# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.GenerateShortURLRequest:
    properties:
      input:
        example: https://www.google.com
        type: string
    type: object
  handler.GenerateShortURLResponse:
    properties:
      short_url:
        example: https://me.li/abc123
        type: string
    type: object
  handler.GetMetadataResponse:
    properties:
      number_of_access:
        type: integer
      original_url:
        type: string
      short_url:
        type: string
    type: object
  utils.BadRequestErrorResponse:
    properties:
      message:
        example: Bad request
        type: string
      status:
        example: 400
        type: integer
    type: object
  utils.ConflictErrorResponse:
    properties:
      message:
        example: Conflict
        type: string
      status:
        example: 409
        type: integer
    type: object
  utils.NotFoundErrorResponse:
    properties:
      message:
        example: URL not found
        type: string
      status:
        example: 404
        type: integer
    type: object
host: localhost:3333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a URL Shortener API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Url Shortener API
  version: "1.0"
paths:
  /shorten:
    post:
      consumes:
      - application/json
      description: Generate a short URL from the input URL
      parameters:
      - description: Input URL
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.GenerateShortURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.GenerateShortURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.BadRequestErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ConflictErrorResponse'
      summary: Generate a short URL
      tags:
      - URL
  /url:
    delete:
      consumes:
      - application/json
      description: Delete a URL entry by providing either the original URL or the
        short URL
      parameters:
      - description: URL to be deleted
        in: query
        name: url
        required: true
        type: string
      - description: Type of URL to be deleted (short_url or original_url)
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.NotFoundErrorResponse'
      summary: Delete a URL entry
      tags:
      - URL
    get:
      consumes:
      - application/json
      description: Get the original url from the shortened url and redirect to it
      parameters:
      - description: URL to get metadata for
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.NotFoundErrorResponse'
      summary: Get a URL entry
      tags:
      - URL
  /url/metadata:
    get:
      consumes:
      - application/json
      description: Get information about any URL entry by providing the original URL
        or the shortened URL
      parameters:
      - description: URL to get metadata for
        in: query
        name: url
        required: true
        type: string
      - description: Type of URL (short_url or original_url)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetMetadataResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.NotFoundErrorResponse'
      summary: Get a URL entry
      tags:
      - URL
swagger: "2.0"